MediaUploadButton: Moving the component to the blocks module
